Migrate grid to POJO
Pretty much ported existing logic and API. We *might* run into
difficulties with loss of reactivity
